API: Rename combobox module to comboboxes
That's because we have two comboboxes now in that module.
