+++ spyder/widgets/comboboxes.py
@@ -24,7 +24,7 @@ from qtpy.QtWidgets import (
     QComboBox, QCompleter, QLineEdit, QSizePolicy, QToolTip)
 
 # Local imports
-from spyder.api.widgets.combobox import SpyderComboBox
+from spyder.api.widgets.comboboxes import SpyderComboBox
 from spyder.config.base import _
 from spyder.py3compat import to_text_string
 from spyder.utils.stylesheet import APP_STYLESHEET
